+++ libs/cdk/data-source/jest.config.ts
@@ -1,9 +1,0 @@
-import baseConfig from '../../../jest.config.base';
-
-export default {
-    ...baseConfig,
-    displayName: 'cdk-data-source',
-    preset: '../../../jest.preset.js',
-    setupFilesAfterEnv: ['<rootDir>/test-setup.ts', '../../../jest-extended-matchers.ts'],
-    coverageDirectory: '../../../dist/coverage/cdk-data-source'
-};