+++ libs/docs/core/notification/examples/notification-mobile/notification-mobile-example.component.ts
@@ -7,7 +7,7 @@ import { MessageToastModule, MessageToastService } from '@fundamental-ngx/core/m
 import { MobileModeConfig } from '@fundamental-ngx/core/mobile-mode';
 import { IndicatorStates, NotificationModule } from '@fundamental-ngx/core/notification';
 import { PopoverComponent, PopoverTriggerDirective } from '@fundamental-ngx/core/popover';
-import { FDP_ICON_TAB_BAR } from '@fundamental-ngx/platform';
+import { FDP_ICON_TAB_BAR } from '@fundamental-ngx/platform/icon-tab-bar';
 
 export type Notification = {
     avatar: string;